chore: add icons for web
