+++ flutter_launcher_icons.yaml
@@ -1,5 +1,7 @@
 flutter_launcher_icons:
   android: "launcher_icon"
   ios: true
+  web:
+    generate: true
   image_path: "assets/app-icon.png"
   min_sdk_android: 21
